requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
python-jose
passlib[bcrypt]
slowapi
pydantic